chore: port over test_yaml_config
