+++ great_expectations/data_context/data_context/abstract_data_context.py
@@ -25,6 +25,12 @@ from typing import (
     cast,
 )
 
+try:
+    from typing import Literal  # type: ignore[attr-defined]
+except ImportError:
+    # Fallback for python < 3.8
+    from typing_extensions import Literal  # type: ignore[misc]
+
 from dateutil.parser import parse
 from ruamel.yaml.comments import CommentedMap
 
@@ -49,6 +55,9 @@ from great_expectations.core.usage_statistics.events import UsageStatsEvents
 from great_expectations.core.util import nested_update
 from great_expectations.core.yaml_handler import YAMLHandler
 from great_expectations.data_asset import DataAsset
+from great_expectations.data_context.config_validator.yaml_config_validator import (
+    _YamlConfigValidator,
+)
 from great_expectations.data_context.data_context_variables import DataContextVariables
 from great_expectations.data_context.store import Store, TupleStoreBackend
 from great_expectations.data_context.store.expectations_store import ExpectationsStore
@@ -195,6 +204,10 @@ class AbstractDataContext(ABC):
 
         # NOTE - 20210112 - Alex Sherstinsky - Validation Operators are planned to be deprecated.
         self.validation_operators: dict = {}
+
+        self._yaml_config_validator = _YamlConfigValidator(
+            data_context=self,
+        )
 
     @abstractmethod
     def _init_variables(self) -> DataContextVariables:
@@ -3074,3 +3087,60 @@ Generated, evaluated, and stored {total_expectations} Expectations during profil
             else:
                 return False
         return include_rendered_content
+
+    def test_yaml_config(  # noqa: C901 - complexity 17
+        self,
+        yaml_config: str,
+        name: Optional[str] = None,
+        class_name: Optional[str] = None,
+        runtime_environment: Optional[dict] = None,
+        pretty_print: bool = True,
+        return_mode: Union[  # type: ignore[name-defined]
+            Literal["instantiated_class"], Literal["report_object"]
+        ] = "instantiated_class",
+        shorten_tracebacks: bool = False,
+    ):
+        """Convenience method for testing yaml configs
+
+        test_yaml_config is a convenience method for configuring the moving
+        parts of a Great Expectations deployment. It allows you to quickly
+        test out configs for system components, especially Datasources,
+        Checkpoints, and Stores.
+
+        For many deployments of Great Expectations, these components (plus
+        Expectations) are the only ones you'll need.
+
+        test_yaml_config is mainly intended for use within notebooks and tests.
+
+        --Public API--
+
+        --Documentation--
+            https://docs.greatexpectations.io/docs/terms/data_context
+            https://docs.greatexpectations.io/docs/guides/validation/checkpoints/how_to_configure_a_new_checkpoint_using_test_yaml_config
+
+        Args:
+            yaml_config: A string containing the yaml config to be tested
+            name: (Optional) A string containing the name of the component to instantiate
+            pretty_print: Determines whether to print human-readable output
+            return_mode: Determines what type of object test_yaml_config will return.
+                Valid modes are "instantiated_class" and "report_object"
+            shorten_tracebacks:If true, catch any errors during instantiation and print only the
+                last element of the traceback stack. This can be helpful for
+                rapid iteration on configs in a notebook, because it can remove
+                the need to scroll up and down a lot.
+
+        Returns:
+            The instantiated component (e.g. a Datasource)
+            OR
+            a json object containing metadata from the component's self_check method.
+            The returned object is determined by return_mode.
+        """
+        return self._yaml_config_validator.test_yaml_config(
+            yaml_config=yaml_config,
+            name=name,
+            class_name=class_name,
+            runtime_environment=runtime_environment,
+            pretty_print=pretty_print,
+            return_mode=return_mode,
+            shorten_tracebacks=shorten_tracebacks,
+        )
